ChangeLog 25
Project submitted.
To be finished at a later date.

+++ CloneBike/Scorer.cpp
@@ -2,7 +2,7 @@
 #include "Libraries.h"
 
 Scorer::Scorer()
-	:Text("0:00:00", Font::ID::Derp, 700)
+	: Text("0:00:00", Font::ID::Derp, 700)
 	, launch(false)
 	, mm(0)
 	, sec(0)
@@ -29,9 +29,11 @@ void Scorer::Update()
 	if (launch)
 	{
 		float dt = Engine::GetInstance()->GetTimer()->GetGameTime();
+		// Intentional, to "shave" off the decimals
 		int temp = dt;
 		float temp2 = dt;
 
+		// Intentional, to remove excess numbers
 		dec = (temp2 - temp) * 100;
 		
 		sec = (int)dt;
@@ -50,7 +52,7 @@ void Scorer::Stop()
 	launch = false;
 }
 
-void Scorer::UpdateTimer(int min, int ss, float dd)
+void Scorer::UpdateTimer(int min, int ss, int dd)
 {
 	this->SetText(std::to_string(mm) + ':' + std::to_string(sec) + ':' + std::to_string(dec));
 }
